requirements/bundle.txt
ankiutils @ git+https://github.com/abdnh/ankiutils@b5c5dfd2a16446781cc2de06fdbd019cfa733c94
    # via -r requirements/bundle.in
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
aw-client==0.5.12
    # via -r requirements/bundle.in
aw-core==0.5.15
    # via aw-client
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via aw-client
colorama==0.4.6
    # via click
deprecation==2.1.0
    # via aw-core
idna==3.4
    # via requests
iso8601==1.1.0
    # via aw-core
jsonschema==4.19.1
    # via aw-core
jsonschema-specifications==2023.7.1
    # via jsonschema
packaging==23.2
    # via deprecation
peewee==3.16.3
    # via aw-core
persist-queue==0.8.1
    # via aw-client
platformdirs==3.11.0
    # via aw-core
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via aw-client
rfc3339-validator==0.1.4
    # via aw-core
rpds-py==0.10.4
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via rfc3339-validator
strict-rfc3339==0.7
    # via aw-core
tabulate==0.9.0
    # via aw-client
timeslot==0.1.2
    # via aw-core
tomlkit==0.12.1
    # via aw-core
typing-extensions==4.8.0
    # via aw-client
urllib3==2.0.6
    # via requests
